interviewRouter refactored to focus solely on defining routes. Controller logic and utility functions are moved it their own modules
