+++ backend/src/api/interview/controllers/index.ts
@@ -1,0 +1,5 @@
+export { extractFacts } from "./extractFacts.js";
+export { getAllInterviews } from "./getAllInterviews.js";
+export { getInterviewById } from "./getInterviewById.js";
+export { transcribeAudio } from "./transcribeAudio.js";
+export { updateAnswer } from "./updateAnswer.js";
