+++ backend/src/api/interview/utils/fileUpload.ts
@@ -1,0 +1,9 @@
+import multer from "multer";
+
+// Configure multer for audio file uploads
+export const upload = multer({
+	storage: multer.memoryStorage(),
+	limits: {
+		fileSize: 10 * 1024 * 1024, // 10MB limit
+	},
+});
