+++ frontend/src/components/pages/InterviewsPage.tsx
@@ -51,7 +51,7 @@ const InterviewsPage = () => {
   };
 
   const formatDate = (dateString: string) => {
-    return new Date(dateString).toLocaleDateString('en-US', {
+    return new Date(dateString).toLocaleDateString('en-UK', {
       year: '2-digit',
       month: '2-digit',
       day: '2-digit'
